Reorganize the basic sources and add a SinWaveSource.
